Move WorkspacePermissionMode inside workspace plugin
