+++ src/core/public/index.ts
@@ -354,4 +354,4 @@ export { __osdBootstrap__ } from './osd_bootstrap';
 
 export { WorkspacesStart, WorkspacesSetup, WorkspacesService } from './workspace';
 
-export { WORKSPACE_TYPE, WorkspacePermissionMode } from '../utils';
+export { WORKSPACE_TYPE } from '../utils';
