+++ src/core/server/index.ts
@@ -360,7 +360,6 @@ export {
   MANAGEMENT_WORKSPACE_ID,
   WORKSPACE_TYPE,
   PERSONAL_WORKSPACE_ID_PREFIX,
-  WorkspacePermissionMode,
 } from '../utils';
 
 export {
