+++ src/core/utils/constants.ts
@@ -7,13 +7,6 @@ export const WORKSPACE_TYPE = 'workspace';
 
 export const WORKSPACE_PATH_PREFIX = '/w';
 
-export enum WorkspacePermissionMode {
-  Read = 'read',
-  Write = 'write',
-  LibraryRead = 'library_read',
-  LibraryWrite = 'library_write',
-}
-
 export const PUBLIC_WORKSPACE_ID = 'public';
 
 export const MANAGEMENT_WORKSPACE_ID = 'management';
