+++ src/plugins/workspace/server/saved_objects/workspace_saved_objects_client_wrapper.ts
@@ -23,7 +23,6 @@ import {
   SavedObjectsBulkUpdateResponse,
   SavedObjectsBulkUpdateOptions,
   WORKSPACE_TYPE,
-  WorkspacePermissionMode,
   SavedObjectsDeleteByWorkspaceOptions,
   SavedObjectsErrorHelpers,
   SavedObjectsServiceStart,
@@ -31,7 +30,10 @@ import {
 } from '../../../../core/server';
 import { SavedObjectsPermissionControlContract } from '../permission_control/client';
 import { getPrincipalsFromRequest } from '../utils';
-import { WORKSPACE_SAVED_OBJECTS_CLIENT_WRAPPER_ID } from '../../common/constants';
+import {
+  WORKSPACE_SAVED_OBJECTS_CLIENT_WRAPPER_ID,
+  WorkspacePermissionMode,
+} from '../../common/constants';
 
 // Can't throw unauthorized for now, the page will be refreshed if unauthorized
 const generateWorkspacePermissionError = () =>
